requirements.txt
django
django-autocomplete-light
django-datetime-widget
django-widget-tweaks